# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fine-Tuning Manager API
  version: 0.1.0
  description: REST API for datasets, evals, fine-tuning, registry, and promotions governed by Orchestrator.
servers:
  - url: https://api.ftm.example.com
security:
  - orchestratorApiKey: []
...
paths:
...
  /ft-jobs:
    post:
      tags: [ft-jobs]
      summary: Create fine-tuning job
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FtJobCreateRequest'
      responses:
        '201':
          description: FT job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FtJob'
        default:
          $ref: '#/components/responses/Error'
  /ft-jobs/{ft_job_id}:
    get:
      tags: [ft-jobs]
      summary: Get fine-tuning job
      parameters:
        - $ref: '#/components/parameters/FtJobId'
      responses:
        '200':
          description: FT job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FtJob'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    orchestratorApiKey:
      type: apiKey
      in: header
      name: Authorization
      description: Use 'Bearer <orchestrator_api_key>'
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
        format: uuid
...
    FtJobId:
      in: path
      name: ft_job_id
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    FtJobCreateRequest:
      type: object
      required: [project_id, method, provider, base_model, dataset_id]
      properties:
        project_id:
          type: string
        method:
          type: string
          enum: [SFT, DPO, RFT]
          enum: [SFT, DPO, RFT]
        provider:
          type: string
          enum: [openai]
        base_model:
          type: string
        dataset_id:
          type: string
        tenant_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        hyperparams:
          type: object
          additionalProperties: true
        cost_threshold_usd:
          type: number
        prism_pre_approval:
          type: boolean
    FtJob:
      type: object
      properties:
        ft_job_id:
          type: string
          format: uuid
        project_id:
          type: string
        provider:
          type: string
        method:
          type: string
        base_model:
          type: string
        dataset_id:
          type: string
        status:
          type: string
          enum: [queued, running, completed, failed, blocked, canceled]
        provider_job_id:
          type: string
        result_json:
          type: object
          additionalProperties: true
        cost_estimate_usd:
          type: number
        cost_actual_usd:
          type: number
        prism_tracked:
          type: boolean
        orchestrator_audit_id:
          type: string
        governance_decision_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        request_id:
          type: string